refactor(upstream): extract pool/logging.rs + pool/entries.rs (lab-kvji.12.3)
Steps 3-4 of the pool.rs split.
- logging.rs: UpstreamRequestLog (+ tool/resource/prompt/with_transport ctors)
  promoted to pub(super) per plan §2.1 (step-3 flip — constructed cross-module
  from tools_call/resources_read/prompts_get in later steps), log_upstream_request_*
  helpers, capability_name, is_capability_unsupported, + the request-log test.
- entries.rs: lazy/healthy/failed entry constructors, failed_from_existing,
  resolve_exposure_policy, health_str, + invalid_exposure_policy test.
- observability_source_* test rewritten to scan the whole pool/ source tree
  (directory-scan) so it stays robust as instrumentation relocates; no assertion
  dropped (guards against silently dropped log fields).
Behavior unchanged. 1423 tests green.

Bead: lab-kvji.12.3
